Move rfc5424 to a new top-level "standards" package.
