+++ standards/doc.go
@@ -1,0 +1,6 @@
+// Copyright 2016 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+// The standards package contains implementations of various industry
+// standards, e.g. protocol specifications in IETF RFCs.
+package protocols
