+++ standards/rfc5424/message-hostname_test.go
@@ -11,7 +11,7 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/logfwd/syslog/rfc5424"
+	"github.com/juju/juju/standards/rfc5424"
 )
 
 type HostnameSuite struct {
