+++ standards/rfc5424/priority.go
