+++ standards/rfc5424/util.go
